Extract PoolContract into a shared component
The pool contract cell now lives in `PoolContract`, which takes
the address as a prop so the upcoming target fixed yield row can
point it at its own contract.
